Fixed the banner overflow as well as separated the Environment into a separate component
