+++ client/src/scss/components/_banner.scss
@@ -2,7 +2,7 @@
 
 .banner {
   background-color: $theme_color_primary;
-  height: 22px;
+  // height: 22px;
   width: 100%;
   font-size: 14px;
 }
