+++ client/src/scss/components/_button.scss
@@ -273,7 +273,8 @@ button {
   color: white;
   font-size: 14px;
   border-radius: 5px;
-  border: 1px white solid;
+  // border: 1px white solid;
+  border: 0px;
   font-family: "phrase_font";
   padding-top: 3px;
 }
